Remove <p/> Elements
I had a few <div/> elements that were nested within <p/> elements. That
is apparently discouraged:
https://stackoverflow.com/questions/41928567/div-cannot-appear-as-a-descendant-of-p

In my use cases, I think the <div/> elements are necessary whereas the
<p/> elements are not, so I'm getting rid of the paragraphs and keeping
the divs instead of the other way around. I'm going to see if there are
jsx linters that are available through npm.

I also did a little bit of configuration in the siteDeploy action due to
some updates to the action I've been using.