Add WIP Counter Widget
This commit adds a small component that I'm calling a "widget". A widget
is a fun component that I want displayed on the front of the website.
Currently the counter widget doesn't do anything, but I will be
iterating on it and, eventually, it is intended to be clickable and
count the number of times it has been clicked.