GuestBook Formatting for Top/Bottom Entry
With this commit, there is special formatting for the first and last entry in the GuestBook. The first/top entry will have rounded edges on the top of the component. Likewise, the bottom edges of the component will be rounded for the last/bottom entry.

This commit also adds the borderRadius to the parent component in the entry row instead of having on the child elements.