Use Jest to Mock useGuestBookInfo
Previously, I created some custom functions to accomplish dependency injection and to provide mock implementations for testing.

It turns out that I can actually accomplish this using Jest fairly easily if I put the logic in a class with methods.

This commit refactors useGuestBookInfo to use a class for its provider implementation so that I can ust jest to mock the implementation in testing. Since they were no longer needed, this commit removes all ProviderRegistry and ProdMode modules and logic.